fix(dependencies): 1. Refactor of classes initialzation
